🔥 Supprime le fichier utils.js de publicodes
