+++ publicodes/source/components/mecanisms/common.tsx
@@ -2,7 +2,7 @@ import React, { createContext, useContext, useState } from 'react'
 import { Trans } from 'react-i18next'
 import styled from 'styled-components'
 import mecanismsDoc from '../../../docs/mecanisms.yaml'
-import { formatValue } from '../../format'
+import { formatValue, capitalise0 } from '../../format'
 import { simplifyNodeUnit } from '../../nodeUnits'
 import { Evaluation, EvaluatedNode, Types, Unit } from '../../AST/types'
 import Overlay from '../Overlay'
@@ -12,7 +12,6 @@ import Explanation from '../Explanation'
 import { ReferenceNode } from '../../reference'
 import { EngineContext } from '../contexts'
 import { InternalError } from '../../error'
-import { capitalise0 } from '../../utils'
 import { Markdown } from '../Markdown'
 
 export function ConstantNode({ nodeValue, type, fullPrecision, unit }) {
