+++ publicodes/source/error.ts
@@ -1,14 +1,12 @@
-import { coerceArray } from './utils'
-
 export class EngineError extends Error {}
 export function syntaxError(
-	rules: string[] | string,
+	rule: string,
 	message: string,
 	originalError?: Error
 ) {
 	throw new EngineError(
 		`\n[ Erreur syntaxique ]
-➡️  Dans la règle \`${coerceArray(rules).slice(-1)[0]}\`
+➡️  Dans la règle \`${rule}\`
 ✖️  ${message}
     ${originalError ? originalError.message : ''}
 `
@@ -16,13 +14,13 @@ export function syntaxError(
 }
 
 export function compilationError(
-	rules: string[] | string,
+	rule: string,
 	message: string,
 	originalError?: Error
 ) {
 	throw new Error(
 		`\n[ Erreur de compilation ]
-➡️ Dans la règle \`${coerceArray(rules).slice(-1)[0]}\`
+➡️ Dans la règle \`${rule}\`
 ✖️ ${message}
   ${originalError?.message}
 `
@@ -30,13 +28,13 @@ export function compilationError(
 }
 
 export function evaluationError(
-	rules: string[] | string,
+	rule: string,
 	message: string,
 	originalError?: Error
 ) {
 	throw new EngineError(
 		`\n[ Erreur d'évaluation ]
-➡️  Dans la règle \`${coerceArray(rules).slice(-1)[0]}\`
+➡️  Dans la règle \`${rule}\`
 ✖️  ${message}
     ${originalError ? originalError.message : ''}
 `
@@ -44,14 +42,14 @@ export function evaluationError(
 }
 
 export function typeWarning(
-	rules: string[] | string,
+	rule: string,
 	message: string,
 	originalError?: Error
 ) {
 	// DESACTIVE EN ATTENDANT L'INFÉRENCE DE TYPE
 	// 	console.warn(
 	// 		`\n[ Erreur de type ]
-	// ➡️  Dans la règle \`${coerceArray(rules).slice(-1)[0]}\`
+	// ➡️  Dans la règle \`${rule}\`
 	// ✖️  ${message}
 	//     ${originalError ? originalError.message : ''}
 	// `
@@ -77,7 +75,7 @@ export class InternalError extends EngineError {
 	constructor(payload) {
 		super(
 			`
-Erreur interne du moteur. 
+Erreur interne du moteur.
 
 Cette erreur est le signe d'un bug dans publicodes. Pour nous aider à le résoudre, vous pouvez copier ce texte dans un nouveau ticket : https://github.com/betagouv/mon-entreprise/issues/new.
